+++ .vscode/settings.json
@@ -1,6 +1,7 @@
 {
     "cSpell.words": [
         "clsx",
-        "Supabase"
+        "Supabase",
+        "tailwindcss"
     ]
 }
