+++ components/Header.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import Image from "next/image";
 import { useState } from "react";
 import { motion } from "framer-motion";
