--- pages/_app.tsx
@@ -1,11 +1,0 @@
-import LayoutWrapper from "@/components/LayoutWrapper";
-import "@/styles/globals.css";
-import type { AppProps } from "next/app";
-
-export default function App({ Component, pageProps }: AppProps) {
-  return (
-    <LayoutWrapper>
-      <Component {...pageProps} />
-    </LayoutWrapper>
-  );
-}
